Margin bottom of zero on last child was causing issue
* need to confirm with @tunde the purpose of the 0 margin
and whether it's ok to remove in all cases or if we need
to target specific use cases (i.e. width where there are
two columns with cards side by side).  The issue (different
heights) is still there when it reduces to a single column
but I think it is done to minimize the space between sections.
If we're ok with the additional space, then this solution
should work.